resolved suggestions
