+++ client/src/components/Main/index.js
@@ -1,6 +1,5 @@
 import React from "react";
 import AppBar from "../Common/AppBar";
-import Map from "../Common/Map";
 import { eventPink } from "../../resources/style-constants";
 
 const Main = () => {
@@ -8,10 +7,6 @@ const Main = () => {
     <div>
       <AppBar color={eventPink} />
       Page coming soon ...
-
-      {/* TODO: Place map in correct place in page. */}
-      <Map />
-
     </div>
   );
 };
